[adsp] Remove AudioDSP from ServiceBroker and ServiceManager (part 4), extract AudioDSP helper methods into new class
